+++ Match/src/Match/vulkan/reource/buffer.cpp
@@ -3,7 +3,7 @@
 #include "../inner.hpp"
 
 namespace Match {
-    Buffer::Buffer(uint32_t size, vk::BufferUsageFlags buffer_usage, VmaMemoryUsage vma_usage, VmaAllocationCreateFlags vma_flags) : size(size), mapped(false), data_ptr(nullptr) {
+    Buffer::Buffer(uint64_t size, vk::BufferUsageFlags buffer_usage, VmaMemoryUsage vma_usage, VmaAllocationCreateFlags vma_flags) : size(size), mapped(false), data_ptr(nullptr) {
         vk::BufferCreateInfo buffer_create_info {};
         buffer_create_info.setUsage(buffer_usage)
             .setSize(size);
@@ -52,7 +52,7 @@ namespace Match {
         vmaDestroyBuffer(manager->vma_allocator, buffer, buffer_allocation);
     }
 
-    TwoStageBuffer::TwoStageBuffer(uint32_t size, vk::BufferUsageFlags usage, vk::BufferUsageFlags additional_usage) {
+    TwoStageBuffer::TwoStageBuffer(uint64_t size, vk::BufferUsageFlags usage, vk::BufferUsageFlags additional_usage) {
         staging = std::make_unique<Buffer>(size, vk::BufferUsageFlagBits::eTransferSrc, VMA_MEMORY_USAGE_CPU_TO_GPU, VMA_ALLOCATION_CREATE_HOST_ACCESS_SEQUENTIAL_WRITE_BIT | VMA_ALLOCATION_CREATE_HOST_ACCESS_ALLOW_TRANSFER_INSTEAD_BIT);
         buffer = std::make_unique<Buffer>(size, usage | vk::BufferUsageFlagBits::eTransferDst | additional_usage, VMA_MEMORY_USAGE_GPU_ONLY, VMA_ALLOCATION_CREATE_DEDICATED_MEMORY_BIT);        
     }
